보스 타격 가능, 플레이어 때리기도 가능 보스전 체력만 넣으면 완성

+++ Team Project/Assets/MyTeam/Scripts/InGame/character.cs
@@ -10,7 +10,7 @@ public class character : MonoBehaviour
     public float      movespeed;                 //움직임 속도 
     public float      criticalpercent;           //크리티컬 확률
     public float      criticaldamage;            //크리티컬 대미지
-    public float      damage;                    //공격력
+    public int      damage;                    //공격력
     public float      attackspeed;               //공격속도
     public float      defence;                   //방어
